Use a common config schema for Dask executor and resource
Combine the Dask config schema used for both the executor and resource.
There is no effective changes for the resource. For the executor, it
will gain client config options.

The executor's "existing" cluster config will be deprecated in favour
of the client config.
