+++ python_modules/libraries/dagster-dask/dagster_dask/resources.py
@@ -1,24 +1,7 @@
-from dagster import Bool, Field, Int, Permissive, Selector, Shape, String, resource
+from dagster import resource
 from dask.distributed import Client
 
-DaskClusterTypes = {
-    key: dict(zip(("name", "module", "class"), values))
-    for key, values in {
-        "local": ("Local", "dask.distributed", "LocalCluster"),
-        "yarn": ("YARN", "dask_yarn", "YarnCluster"),
-        "ssh": ("SSH", "dask.distributed", "SSHCluster"),
-        "pbs": ("PBS", "dask_jobqueue", "PBSCluster"),
-        "moab": ("Moab", "dask_jobqueue", "MoabCluster"),
-        "sge": ("SGE", "dask_jobqueue", "SGECluster"),
-        "lsf": ("LSF", "dask_jobqueue", "LSFCluster"),
-        "slurm": ("SLURM", "dask_jobqueue", "SLURMCluster"),
-        "oar": ("OAR", "dask_jobqueue", "OARCluster"),
-        "kube": ("Kubernetes", "dask_kubernetes", "KubeCluster"),
-        "azureml": ("Azure ML", "dask_cloudprovider", "AzureMLCluster"),
-        "ecs": ("AWS ECS", "dask_cloudprovider", "ECSCluster"),
-        "fargate": ("AWS ECS on Fargate", "dask_cloudprovider", "FargateCluster"),
-    }.items()
-}
+from .config import DaskConfigSchema
 
 
 class DaskResource:
@@ -67,77 +50,7 @@ class DaskResource:
 
 @resource(
     description="Dask Client resource.",
-    config_schema=Shape(
-        {
-            "client": Field(
-                Permissive(
-                    {
-                        "address": Field(
-                            String,
-                            description="Address of a Scheduler server.",
-                            is_required=False,
-                        ),
-                        "timeout": Field(
-                            Int,
-                            description="Timeout duration for initial connection to the scheduler.",
-                            is_required=False,
-                        ),
-                        "set_as_default": Field(
-                            Bool,
-                            description="Claim this scheduler as the global dask scheduler.",
-                            is_required=False,
-                        ),
-                        "scheduler_file": Field(
-                            String,
-                            description="Path to a file with scheduler information, if available.",
-                            is_required=False,
-                        ),
-                        "security": Field(
-                            Bool,
-                            description="Optional security information.",
-                            is_required=False,
-                        ),
-                        "asynchronous": Field(
-                            Bool,
-                            description="Set to True if using this client within async/await functions.",
-                            is_required=False,
-                        ),
-                        "name": Field(
-                            String,
-                            description="Name of client that will be included in logs generated on the scheduler.",
-                            is_required=False,
-                        ),
-                        "direct_to_workers": Field(
-                            Bool,
-                            description="Whether to connect directly to the workers.",
-                            is_required=False,
-                        ),
-                        "heartbeat_interval": Field(
-                            Int,
-                            description="Time in milliseconds between heartbeats to scheduler.",
-                            is_required=False,
-                        ),
-                    }
-                ),
-                description="Dask distributed client options.",
-                is_required=False,
-            ),
-            "cluster": Field(
-                Selector(
-                    {
-                        key: Field(
-                            Permissive(),
-                            is_required=False,
-                            description=f"{meta['name']} cluster config. Requires {meta['module']}.",
-                        )
-                        for key, meta in DaskClusterTypes.items()
-                    }
-                ),
-                description="Create a Dask cluster. Will be passed as the client address.",
-                is_required=False,
-            ),
-        }
-    ),
+    config_schema=DaskConfigSchema,
 )
 def dask_resource(context):
     res = DaskResource(context)
